// frontend/app/shared/infrastructure/db/model/kysely/tables.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
	? ColumnType<S, I | undefined, U>
	: ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<
	string,
	bigint | number | string,
	bigint | number | string
>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
	[x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Account {
	email: string;
	emailVerified: Generated<boolean>;
	name: string;
	passwordHash: string;
	preferences: Generated<Json>;
	uuid: Generated<string>;
}
// ...
export interface Customer {
	address: string | null;
	city: string | null;
	company: string | null;
	country: string | null;
	customerId: Generated<number>;
	email: string;
	fax: string | null;
	firstName: string;
	lastName: string;
	phone: string | null;
	postalCode: string | null;
	state: string | null;
	supportRepId: number | null;
}
// ...
export interface Netflix {
	casts: string | null;
	country: string | null;
	dateAdded: string | null;
	description: string | null;
	director: string | null;
	duration: string | null;
	globalText: Generated<string | null>;
	listedIn: string | null;
	rating: string | null;
	releaseYear: number | null;
	showId: string | null;
	title: string | null;
	type: string | null;
}